+++ src/IMDb/Controllers/loadingController.java
@@ -1,9 +1,10 @@
-package IMDb;
+package IMDb.Controllers;
+
+import IMDb.Classes.*;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.Initializable;
-import javafx.scene.control.Button;
 import javafx.scene.control.ProgressBar;
 
 import java.util.concurrent.TimeUnit;
@@ -11,7 +12,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class Controller implements Initializable {
+public class loadingController implements Initializable {
 
    public ProgressBar progressBar;
 
@@ -41,7 +42,7 @@ public class Controller implements Initializable {
         done = 0.0;
         size = 0.0;
         ObservableList<movie> movies = FXCollections.observableArrayList();
-        database data = new database("C:/Users/jeffr/Documents/ICS4U1/CPT/ics4u-cpt---data-visualization-jeffrey-lin/src/IMDb/dataset_full.csv");
+        database data = new database("src/IMDb/Resources/dataset_full.csv");
         size = data.getSize() * 1.0;
             
         for(String id: data.getMovieList()) {
