--- src/IMDb/loadingScreen.fxml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.text.*?>
-<?import javafx.scene.shape.*?>
-<?import javafx.geometry.*?>
-<?import javafx.scene.image.*?>
-<?import java.lang.*?>
-<?import javafx.scene.control.*?>
-<?import javafx.scene.layout.*?>
-
-<VBox alignment="CENTER" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" style="-fx-background-color: #262626;" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="IMDb.Controller">
-   
-   <children>
-      <ImageView fitHeight="277.0" fitWidth="415.0" layoutX="93.0" layoutY="84.0" nodeOrientation="INHERIT" pickOnBounds="true" preserveRatio="true">
-         <image>
-            <Image url="@logo.png" />
-         </image>
-         <VBox.margin>
-            <Insets />
-         </VBox.margin>
-      </ImageView>
-      <ProgressBar fx:id="progressBar" layoutX="92.0" layoutY="293.0" minHeight="0.0" prefHeight="8.0" prefWidth="419.0" progress="0.0" style="-fx-text-box-border: #262626; -fx-control-inner-background: #262626; -fx-background-radius: 0;">
-         <VBox.margin>
-            <Insets top="5.0" />
-         </VBox.margin></ProgressBar>
-      <Rectangle arcHeight="5.0" arcWidth="5.0" fill="#262626" height="34.0" layoutX="77.0" layoutY="297.0" stroke="#262626" strokeType="INSIDE" width="452.0">
-         <VBox.margin>
-            <Insets top="-3.0" />
-         </VBox.margin></Rectangle>
-   </children>
-</VBox>
