+++ src/IMDb/main.java
@@ -15,11 +15,12 @@ public class main extends Application {
 
     @Override
     public void start(Stage primaryStage) throws Exception{
-        Parent root = FXMLLoader.load(getClass().getResource("loadingScreen.fxml"));
+        Parent root = FXMLLoader.load(getClass().getResource("FXML/loadingScreen.fxml"));
         Scene loadingScreen = new Scene(root, Screen.getPrimary().getBounds().getMaxX()/2,Screen.getPrimary().getBounds().getMaxY()/2);
         primaryStage.setTitle("Jeffrey's Movie Database");
         primaryStage.setScene(loadingScreen);
         primaryStage.show();
+        //primaryStage.close();
     }
